requirements.txt
# Django Framework
Django==5.2.7

# Database
psycopg2-binary==2.9.9

# Environment Variables
python-decouple==3.8
django-environ==0.11.2

# Static Files
whitenoise==6.6.0

# Production Server
gunicorn==21.2.0

# Image Processing
Pillow==10.1.0

# Timezone
pytz==2023.3

# Additional Production Dependencies
django-cors-headers==4.3.1
django-extensions==3.2.3

# HTTPS للتطوير (حل ERR_SSL_PROTOCOL_ERROR)
Werkzeug==3.0.4
pyOpenSSL==24.1.0
daphne==4.1.3

# Excel Export
openpyxl==3.1.2